rpi/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pyserial==3.5
python-dotenv==1.0.0
